2-10

+++ Practice/src/practice02/PTra02_09.java
@@ -18,7 +18,7 @@ public class PTra02_09 {
 				System.out.println(quadrangle);
 
 		// ★ 変数triangleに、三角形の面積を代入してください
-			double triangle=length*height/2;
+			double triangle=(double)length*height/2;
 
 		// ★ 変数triangleを出力してください
 						System.out.println(triangle);
